* develop:
  Finished first time log in functionality
  Improved auth UI
  Added delete user UI
  Added topic functionality.
  Added code to handle first time log in (unfinished)
  fixed notifier styling links
  improved login form validation
  added UI response feedback to log in form
  Use API to populate users list
  Fixed /users and /users/{id} endpoints
  Added 'add' and 'edit' modals
  Added users page layout
  Added get all queue items functionality, including the response, queue dto and queue controller
  Re-added get timespan and get timespans code with other changes
  Cleaned up authentication.service syntax
  Added logout button
  Implemented AuthGuard and return URL redirect

# Conflicts:
#	HelpdeskSystem/Helpdesk.Services.Test/QueueTests.cs

+++ HelpdeskSystem/Helpdesk.Services/UsersFacade.cs
@@ -136,7 +136,9 @@ namespace Helpdesk.Services
 
                 if (dataLayer.GetUserByUsername(request.Username) != null)
                 {
-                    throw new Exception("Unable to add user! User already exists!");
+                    response.Status = HttpStatusCode.BadRequest;
+                    response.StatusMessages.Add(new StatusMessage(HttpStatusCode.BadRequest, "Username already exists"));
+                    return response;
                 }
 
                 int? result = dataLayer.AddUser(request);
@@ -183,7 +185,7 @@ namespace Helpdesk.Services
 
                 var dataLayer = new UsersDataLayer();
 
-                if (dataLayer.GetUserByUsername(request.Username) != null)
+                if (dataLayer.GetUserByUsername(request.Username).UserId != id)
                 {
                     throw new Exception("Unable to update user! User with username " + request.Username + "already exists!");
                 }
@@ -217,13 +219,22 @@ namespace Helpdesk.Services
         /// </summary>
         /// <param name="id">The id of the user to be deleted</param>
         /// <returns>A response that indicates whether or not the deletion was successful</returns>
-        public DeleteUserResponse DeleteUser(int id)
+        public DeleteUserResponse DeleteUser(int id, string currentUser)
         {
             var response = new DeleteUserResponse();
 
             try
             {
                 var dataLayer = new UsersDataLayer();
+
+                UserDTO user = dataLayer.GetUser(id);
+
+                if (user.Username == currentUser)
+                {
+                    response.Status = HttpStatusCode.Forbidden;
+                    return response;
+                }
+
                 bool result = dataLayer.DeleteUser(id);
 
                 if (result)
@@ -284,6 +295,13 @@ namespace Helpdesk.Services
                     response.Status = HttpStatusCode.BadRequest;
                     response.StatusMessages.Add(new StatusMessage(HttpStatusCode.BadRequest, "Unable to login username or password is incorrect"));
                     s_logger.Warn($"Unable to login as a username [ {request.Username} ], username or password is incorrect.");
+                    return response;
+                }
+
+                if (user.FirstTime)
+                {
+                    response.Status = HttpStatusCode.Accepted;
+                    response.UserId = user.UserId;
                     return response;
                 }
